fix structure and namings
