+++ src/routes/v1/api.js
@@ -8,7 +8,7 @@ import authRouter from "./auth.js";
 import dashboardRouter from "./dashboard.js";
 import governmentRouter from "./governments.js";
 import specializationRouter from "./specializations.js";
-import chatRouter from "../chat.js";
+import chatRouter from "./chat.js";
 import { isActive, authenticateAccess } from "../../middlewares/authMiddleware.js";
 import { sensitiveIpRateLimiter } from "../../middlewares/rateLimitMiddleware.js";
